moved class FockQuantumChannel to example/quantum_transduction/fock_quantum_channel.py
